+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 * Fix: Visual Studio locks Reqnroll.Tools.MsBuild.Generation task files. Using `TaskHostFactory` for our tasks on Windows. (#293)
 * Fix: Project dependencies transiently refer to System.Text.Json 8.0.4 that has security vulnerability. Microsoft.Extensions.DependencyModel updated to v8.0.2. (#291)
+* Fix: Could not load System.CodeDom exception with xRetry.Reqnroll plugin (#310)
 
 *Contributors of this release (in alphabetical order):* @gasparnagy, @obligaron, @Romfos, @Tiberriver256